+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.73.0 (Unreleased)
 ENHANCEMENTS:
 * compute: add `host_affinity_rules` attribute to `placement_policy` in `yandex_compute_instance` resource and data source
+* alb: support `storage_bucket` as a target in `yandex_alb_backend_group` resource and data source
 
 BUG FIXES:
 * compute: avoid unnecessary start/stop operations for instance while updating secondary disk configuration